Move flow expander to components from modules.
