+++ packages/explorable-explanations/src/protectedAudience/components/index.js
@@ -17,3 +17,4 @@ export { default as Box } from './box';
 export { default as Branches } from './branches';
 export { default as ProgressLine } from './progressLine';
 export { default as RippleEffect } from './rippleEffect';
+export { default as FlowExpander } from './flowExpander';
